requirements.txt
flask
flask-sqlalchemy
azure-cognitiveservices-speech
pydub
google-generativeai